pyproject.toml
[project]
name = "softgear"
version = "0.1.0"
description = "Learnable Asymmetric Gear Chains for Structured Reasoning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "pandas>=3.0.1",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "wandb>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "huggingface-hub>=1.5.0",
    "pytest>=9.0.2",
    "pytest-timeout>=2.4.0",
]
